# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
...
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        metadata:
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - status
        - version
        - createdAt
...
    MarketComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        marketId:
          type: string
        authorId:
          type: string
          nullable: true
          format: uuid
        authorAddress:
          type: string
          nullable: true
        body:
          type: string
        moderationFlagged:
          type: boolean
        moderationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - marketId
        - authorId
        - authorAddress
        - body
        - moderationFlagged
        - moderationReason
        - createdAt
...
    FeatureFlagListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FeatureFlag'
      required:
        - data
...
paths:
...
  /api/markets/recommendations:
    get:
      operationId: getMarketRecommendations
      tags:
        - Markets
      summary: Get personalized market recommendations
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            maxLength: 512
          required: false
          name: cursor
          in: query
      responses:
        '200':
          description: Paginated list of recommended markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
              examples:
                recommendedMarkets:
                  value:
                    data:
                      - id: market-003
                        question: Will Bitcoin close above $100k in 2026?
                        status: active
                        metadata:
                          category: crypto
                          resolutionSource: official
                        version: 1
                        createdAt: '2026-03-01T09:00:00.000Z'
                    nextCursor: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                unauthorized:
                  value:
                    error:
                      code: unauthorized
                      requestId: req_abc123
  /api/recommendations:
    get:
      operationId: getRecommendations
      tags:
        - Markets
      summary: Get personalized market recommendations
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            maxLength: 512
          required: false
          name: cursor
          in: query
      responses:
        '200':
          description: Paginated list of recommended markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/markets:
    get:
      operationId: listMarkets
      tags:
        - Markets
      summary: List all markets with cursor pagination
      description: >-
        Returns a cursor-paginated list of markets. Supports strong ETag / conditional GET: send the ETag back as
        If-None-Match on subsequent requests; if the page is unchanged the server responds 304 Not Modified (no body).
      parameters:
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when the page is unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: Array of markets
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                required:
                  - data
              examples:
                default:
                  value:
                    data:
                      - id: market-001
                        question: Will the US win the 2026 FIFA World Cup?
                        status: active
                        metadata:
                          category: sports
                          resolutionSource: official
                        version: 1
                        createdAt: '2026-01-10T12:00:00.000Z'
                      - id: market-002
                        question: Will Stellar launch a new protocol upgrade in 2026?
                        status: active
                        metadata:
                          category: technology
                          resolutionSource: community
                        version: 2
                        createdAt: '2026-02-14T07:30:00.000Z'
        '304':
          description: Not Modified — page unchanged since the ETag in If-None-Match.
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/markets/search:
    get:
      operationId: searchMarkets
      tags:
        - Markets
      summary: Full-text search across markets
      description: >-
        Full-text search with fuzzy trigram fallback. Supports strong ETag / conditional GET: send the ETag back as
        If-None-Match; if results are unchanged the server responds 304 Not Modified.
      parameters:
        - schema:
            type: string
            minLength: 1
          required: true
          name: q
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: page
          in: query
        - schema:
            type: string
            description: ETag from a previous 200 response. Triggers 304 when results are unchanged.
          required: false
          name: If-None-Match
          in: header
      responses:
        '200':
          description: Search results
          headers:
            ETag:
              description: Strong ETag (SHA-256) of the response body.
              schema:
                type: string
            Cache-Control:
              description: Always no-cache so clients revalidate before reuse.
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSearchResult'
              examples:
                searchResults:
                  value:
                    data:
                      - id: market-001
                        question: Will the US win the 2026 FIFA World Cup?
                        status: active
                        metadata:
                          category: sports
                          resolutionSource: official
                        version: 1
                        createdAt: '2026-01-10T12:00:00.000Z'
                    total: 1
                    limit: 20
                    offset: 0
                    page: 1
                    fallback: false
                    pagination:
                      limit: 20
                      offset: 0
                      page: 1
                      total: 1
                      fallback: false
                    meta:
                      limit: 20
                      offset: 0
                      page: 1
                      total: 1
                      fallback: false
        '304':
          description: Not Modified — search results unchanged since the ETag in If-None-Match.
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                missingQuery:
                  value:
                    error:
                      code: validation_error
                      requestId: req_abc123
...
  /api/comments:
    get:
      operationId: listComments
      tags:
        - Comments
      summary: List comments (root endpoint)
      description: Returns a paginated list of comments. The resolved `X-Correlation-Id` is echoed back in the response header.
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            maximum: 100
          required: false
          name: limit
          in: query
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: string
            description: >-
              Client-supplied correlation ID. Alphanumeric, hyphens, and underscores only (max 128 chars). A UUID v4 is
              generated when absent.
          required: false
          name: x-correlation-id
          in: header
      responses:
        '200':
          description: Paginated comments list
          headers:
            x-correlation-id:
              schema:
                type: string
                description: Resolved correlation ID, echoed back to the caller.
              required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketComment'
                  nextCursor:
                    type: string
                    nullable: true
                  message:
                    type: string
                required:
                  - data
                  - nextCursor
                  - message
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
    post:
      operationId: createComment
      tags:
        - Comments
      summary: Create a comment
      description: >-
        Creates a new comment. When `outboundUrl` is provided the service posts the comment payload to that URL,
        forwarding `X-Correlation-Id` so the receiving system can correlate the call.
      parameters:
        - schema:
            type: string
            description: Client-supplied correlation ID propagated to the outbound webhook call.
          required: false
          name: x-correlation-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          headers:
            x-correlation-id:
              schema:
                type: string
                description: Resolved correlation ID, echoed back to the caller.
              required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCommentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
...
  /api/leaderboard:
    get:
      operationId: getLeaderboard
      tags:
        - Leaderboard
      summary: Get global leaderboard
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            maximum: 100
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: boolean
            nullable: true
            default: false
          required: false
          name: refresh
          in: query
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  meta:
                    type: object
                    properties:
                      limit:
                        type: number
                      offset:
                        type: number
                      count:
                        type: number
                      refresh:
                        type: boolean
                    required:
                      - limit
                      - offset
                      - count
                      - refresh
                required:
                  - data
                  - meta
...
  /api/rate-limit:
    get:
      operationId: listRateLimitEvents
      tags:
        - Rate Limiting
      summary: List rate-limit audit events (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated rate-limit audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RateLimitAuditEntry'
                  nextCursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - nextCursor
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/feature-flags:
    get:
      operationId: listAdminFeatureFlags
      tags:
        - Admin
      summary: List configured feature flags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of feature flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    post:
      operationId: createAdminFeatureFlag
      tags:
        - Admin
      summary: Create a new feature flag
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureFlagRequest'
      responses:
        '201':
          description: Feature flag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Feature flag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/notifications/broadcast:
    post:
      operationId: adminBroadcastNotification
      tags:
        - Admin
        - Notifications
      summary: Broadcast notification to all users (admin only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastNotificationRequest'
      responses:
        '201':
          description: Notification successfully broadcast to users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BroadcastNotificationResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '403':
          description: Forbidden — caller is not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin:
    get:
      operationId: listAdminEndpoints
      tags:
        - Admin
      summary: List available admin endpoints
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
          required: false
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Paginated admin endpoint catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...